Fix the return type on 64bit.
Returning unsigned int instead of usigned long will avoid giving the
impression that we return 64bits of sha1 on 64 bit platforms

+++ sha1trunc.h
@@ -291,7 +291,7 @@ static void SHA1_Final(SHA1_CTX* context, uint8_t digest[SHA1_DIGEST_SIZE])
 #endif
 }
 
-static inline unsigned long sum(void *data, int len)
+static inline unsigned int sum(void *data, int len)
 {
     SHA1_CTX sha;
     uint8_t digest[20];
